Update pdf-processing.sh
Now checks if a PDF passed to the script is valid. If it's not, asks if you want to continue anyways.

For prior versions, now extracts # of versions-1, as the last one is the original PDF being processed. It was redundant to re-process that one for images.

The script now also checks if a prior version is a valid PDF and notes that on screen and in the file. If it's not, it does not attempt to execute the pdfImages function on that version.