Fixed extractions, added test data
The previous script extracted a false positive for the first version. Corrected that and some of the text in the output.

Also included a test PDF that has 4 prior edits done to it. The text file "Edits.txt" contains a summary of the edits. And a subfolder contains a copy of each edit as it was done for comparison with extracted versions.