exiftool of all versions
With the update to the script not extracting the current version as a "prior" version, the current version exif data was no longer in the "all versions" csv file. It was only in its own CSV. That was not convenient for comparing exif from all versions including the current.

Script now includes the current version in the "all versions" csv.