Now also extracts # of fonts
PDFs that are non-searchable will have 0 fonts. But searchable PDFs will have fonts. A larger than normal # of fonts can be a potential red flag. Note that a regular font and a bold of that font is 2 fonts. An italic of that same font comes up as 3 fonts. This is simply an additional data point to help triage a large volume of PDFs for those you want to target first based on thresholds you decide to use to flag suspicious PDFs.