Updated pdfimages function
The updated script extracts the images a second time in PNG format, as some images are in .ccitt format when extracting in native format which is not readily supported. That extraction is preserved to keep images in source format. But extracted a second time in PNG to facilitate viewing the images for comparision.